meeting-scheduler/config/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pytz==2023.3
python-dateutil==2.8.2